requirements.txt
matplotlib
opencv-python
torch
torchvision
mmengine
wandb
numpy==1.23.1
pillow
imageio==2.19.3
imageio-ffmpeg==0.4.7
decord
einops==0.4.1
easydict
tqdm
OpenEXR==3.3.1
h5py
pandas
scipy
wand
huggingface_hub
fvcore